requirements.txt
langgraph>=0.0.60
langchain-core>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
fastmcp 